Refactor: move class to new utils package
