+++ p2p/src/main/java/bisq/network/utils/Utils.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.network.p2p;
+package bisq.network.utils;
 
 import java.net.ServerSocket;
